cleanup comments, and move interfaces to external_interfaces.go
