+++ internal/server/server.go
@@ -24,25 +24,6 @@ import (
 //go:embed templates
 var embeddedFS embed.FS
 
-type OIDCClientProvider interface {
-	NewProvider(ctx context.Context, issuer string) (*oidc.Provider, error)
-}
-
-type OIDCClient struct{}
-
-func (o *OIDCClient) NewProvider(ctx context.Context, issuer string) (*oidc.Provider, error) {
-	return oidc.NewProvider(ctx, issuer)
-}
-
-type Oauth2ConfigProvider interface {
-	AuthCodeURL(state string, opts ...oauth2.AuthCodeOption) string
-	Exchange(ctx context.Context, code string, opts ...oauth2.AuthCodeOption) (*oauth2.Token, error)
-}
-
-type OIDCIDTokenVerifier interface {
-	Verify(ctx context.Context, rawIDToken string) (*oidc.IDToken, error)
-}
-
 type Server struct {
 	*mux.Router
 	oidcProvider        OIDCClientProvider
@@ -195,7 +176,7 @@ func (s *Server) handleLogin() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		session := s.getSession(r)
 
-		// Normally this randomStr would actually be a state, and saved in session.Values["state"].
+		// Normally this randomStr would be a state, and saved in session.Values["state"].
 		// Then in handleCallback() check the state in URL matches the session.Values["state"].
 		// However, our app does not do the full login -> redirect -> callback loop.
 		// The user hits /login, copies code, goes back to root page, and hits /callback.
